+++ tests/phpunit/REST/BulkControllerGroupMatchesTest.php
@@ -98,4 +98,72 @@ class BulkControllerGroupMatchesTest extends TestCase {
 		$this->assertSame( 10, $articles[0]['post_id'] );
 		$this->assertSame( 11, $articles[1]['post_id'] );
 	}
+
+	/**
+	 * Carousel leftovers for a heading that already has an image leave Review.
+	 *
+	 * @return void
+	 */
+	public function test_omit_inserted_headings_drops_pending_carousel_rows(): void {
+		$hash = \SmartImageMatcher\Insertion\HeadingLocator::computeHash( 2, 'bald eagle', 0 );
+		$post = new \WP_Post();
+		$post->ID           = 284568;
+		$post->post_content = '<!-- wp:heading --><h2>Bald Eagle</h2><!-- /wp:heading -->';
+		$GLOBALS['sim_test_get_post'] = static function () use ( $post ) {
+			return $post;
+		};
+		$GLOBALS['sim_test_parse_blocks'] = static function () {
+			return array(
+				array(
+					'blockName'   => 'core/heading',
+					'innerHTML'   => '<h2>Bald Eagle</h2>',
+					'attrs'       => array( 'level' => 2 ),
+					'innerBlocks' => array(),
+				),
+				array(
+					'blockName'   => 'core/image',
+					'innerHTML'   => '<figure><img src="bald-eagle.jpg" /></figure>',
+					'attrs'       => array( 'id' => 1 ),
+					'innerBlocks' => array(),
+				),
+			);
+		};
+
+		$kept = BulkController::omitInsertedHeadings(
+			array(
+				array(
+					'id'               => 1,
+					'post_id'          => 284568,
+					'heading_tag'      => 'h2',
+					'heading_text'     => 'Bald Eagle',
+					'heading_hash'     => $hash,
+					'image_id'         => 10,
+					'confidence_score' => 100,
+				),
+				array(
+					'id'               => 2,
+					'post_id'          => 284568,
+					'heading_tag'      => 'h2',
+					'heading_text'     => 'Bald Eagle',
+					'heading_hash'     => $hash,
+					'image_id'         => 11,
+					'confidence_score' => 100,
+				),
+				array(
+					'id'               => 3,
+					'post_id'          => 284568,
+					'heading_tag'      => 'h2',
+					'heading_text'     => 'Golden Eagle',
+					'heading_hash'     => 'hash-still-open',
+					'image_id'         => 12,
+					'confidence_score' => 80,
+				),
+			)
+		);
+
+		unset( $GLOBALS['sim_test_parse_blocks'], $GLOBALS['sim_test_get_post'] );
+
+		$this->assertCount( 1, $kept );
+		$this->assertSame( 'hash-still-open', $kept[0]['heading_hash'] );
+	}
 }